dnekcab-eciovni/requirements/local.txt
# Django
# ------------------------------------------------------------------------------
django==2.0.8  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.4  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.36.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis
django-dirtyfields==1.3.1
django-extensions==2.0.7  # https://github.com/django-extensions/django-extensions
django-cors-headers==2.4.0

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-filter==2.0.0

Werkzeug==0.14.1  # https://github.com/pallets/werkzeug

ipdb==0.11  # https://github.com/gotcha/ipdb
Sphinx==1.7.5  # https://github.com/sphinx-doc/sphinx
psycopg2==2.7.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
factory-boy==2.11.1  # https://github.com/FactoryBoy/factory_boy
django-test-plus==1.1.1  # https://github.com/revsys/django-test-plus

django-debug-toolbar==1.9.1  # https://github.com/jazzband/django-debug-toolbar
django-coverage-plugin==1.5.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.3.2  # https://github.com/pytest-dev/pytest-django

pytest==3.6.2  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.1  # https://github.com/Frozenball/pytest-sugar
pytest-xdist==1.20.1
pytest-cov==2.5.1

# Code quality
# ------------------------------------------------------------------------------
flake8==3.5.0  # https://github.com/PyCQA/flake8
coverage==4.5.1  # https://github.com/nedbat/coveragepy

# Project
# ------------------------------------------------------------------------------
model-mommy==1.5.0
isort>=4.2.15
# Python-Mysql Database Client
mysqlclient==1.3.12
# Django Simple History
django-simple-history==2.3.0
